% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgscript.R
\name{load_run}
\alias{load_run}
\title{Load a simulation run from file.}
\usage{
load_run(file, text = NULL)
}
\arguments{
\item{file}{name of file containing \code{yaml} run specification}

\item{text}{yaml code specifying a simultion run}
}
\description{
Load a simulation run from file.
}
