% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgscript.R
\name{sim_run}
\alias{sim_run}
\title{Simulate from a scripted run.}
\usage{
sim_run(mod, x, .Request = x$endpoints)
}
\arguments{
\item{mod}{a model object}

\item{x}{the simrun object}

\item{.Request}{requested endpoints}
}
\description{
Simulate from a scripted run.
}
\details{
The object passed as \code{x} is the
result of a call to \code{\link{load_run}}.
}
