% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgscript.R
\name{load_run}
\alias{load_run}
\title{Load a simulation run from file.}
\usage{
load_run(file)
}
\arguments{
\item{file}{the name of a file containing \code{yaml} run specification}
}
\description{
Load a simulation run from file.
}
